Email Service implemented

+++ src/tasks/controller.py
@@ -20,8 +20,8 @@ def create_task(body: TaskSchema, db:Session, user:UserModel):
 
     return new_task
 
-def get_task(db:Session):
-    tasks = db.query(TaskModel).all()
+def get_task(db:Session, user:UserModel):
+    tasks = db.query(TaskModel).filter(TaskModel.user_id == user.id).all()
     return tasks
 
 def get_one_task(task_id:int, db:Session):
@@ -32,11 +32,14 @@ def get_one_task(task_id:int, db:Session):
     return one_task
 
 
-def update_task(body:TaskSchema, task_id:int, db:Session):
-    one_task = db.query(TaskModel).get(task_id)
+def update_task(body:TaskSchema, task_id:int, db:Session, user:UserModel):
+    one_task:TaskModel = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404, detail="Task Id  is Incorrect")
     
+    if one_task.user_id != user.id:
+        raise HTTPException(401, detail="You are not allowed to update this task")
+
     body = body.model_dump()
     for field, value in body.items():
         setattr(one_task, field, value)
@@ -48,11 +51,16 @@ def update_task(body:TaskSchema, task_id:int, db:Session):
     return one_task
 
 
-def delete_task(task_id:int, db:Session):
+def delete_task(task_id:int, db:Session, user:UserModel):
     one_task = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404, detail="Task Id  is Incorrect")
     
+
+    if one_task.user_id != user.id:
+        raise HTTPException(401, detail="You are not allowed to delete this task")
+
+
     db.delete(one_task)
     db.commit()
